requirements.txt
fastapi>=0.115.5
hypercorn>=0.14.4
pydantic>=2.10.1
pydantic-settings>=2.5.2
litellm>=1.54.1
opensearch-py>=2.8.0
pinecone-client>=5.4.2
langfuse>=2.57.0
python-json-logger>=2.0.7
pytz>=2024.2
python-dotenv>=1.0.0
